Updates subnet client command for relocated network logic.
