+++ cmd/juju/subnet/subnet.go
@@ -16,7 +16,7 @@ import (
 	"github.com/juju/juju/api/subnets"
 	"github.com/juju/juju/apiserver/params"
 	"github.com/juju/juju/cmd/modelcmd"
-	"github.com/juju/juju/network"
+	"github.com/juju/juju/core/network"
 )
 
 // SubnetAPI defines the necessary API methods needed by the subnet
